move travel certificate usecases and graphql to TC module
