--- app/data/data-travel-certificate/data-travel-certificate-public/build.gradle.kts
@@ -1,38 +1,0 @@
-plugins {
-  id("hedvig.android.ktlint")
-  id("hedvig.android.library")
-  alias(libs.plugins.squareSortDependencies)
-  alias(libs.plugins.apollo)
-}
-
-dependencies {
-  apolloMetadata(projects.apolloOctopusPublic)
-  implementation(libs.apollo.runtime)
-  implementation(libs.apollo.normalizedCache)
-  implementation(libs.arrow.core)
-  implementation(libs.koin.core)
-  implementation(projects.apolloCore)
-  implementation(projects.apolloOctopusPublic)
-  implementation(projects.coreCommonPublic)
-  implementation(projects.featureFlagsPublic)
-  implementation(projects.dataContractPublic)
-
-  testImplementation(libs.apollo.testingSupport)
-  testImplementation(libs.assertK)
-  testImplementation(libs.coroutines.test)
-  testImplementation(libs.junit)
-  testImplementation(libs.testParameterInjector)
-  testImplementation(libs.turbine)
-  testImplementation(projects.apolloOctopusTest)
-  testImplementation(projects.apolloTest)
-  testImplementation(projects.coreCommonTest)
-  testImplementation(projects.featureFlagsTest)
-  testImplementation(projects.loggingTest)
-
-}
-
-apollo {
-  service("octopus") {
-    packageName.set("octopus")
-  }
-}
