--- app/data/data-travel-certificate/data-travel-certificate-test/build.gradle.kts
@@ -1,12 +1,0 @@
-plugins {
-  id("hedvig.android.ktlint")
-  id("hedvig.android.library")
-  alias(libs.plugins.squareSortDependencies)
-}
-
-dependencies {
-  implementation(libs.arrow.core)
-  implementation(libs.turbine)
-  implementation(projects.coreCommonPublic)
-  implementation(projects.dataTravelCertificatePublic)
-}
