--- app/data/data-travel-certificate/data-travel-certificate-test/src/main/kotlin/com/hedvig/android/data/travelcertificate/test/TestGetTravelCertificateSpecificationsUseCase.kt
@@ -1,12 +1,0 @@
-package com.hedvig.android.data.travelcertificate
-
-import app.cash.turbine.Turbine
-import arrow.core.Either
-
-class TestGetTravelCertificateSpecificationsUseCase : GetTravelCertificateSpecificationsUseCase {
-  val turbine = Turbine<Either<TravelCertificateError, TravelCertificateData>>()
-
-  override suspend fun invoke(contractId: String?): Either<TravelCertificateError, TravelCertificateData> {
-    return turbine.awaitItem()
-  }
-}
