+++ app/feature/feature-home/build.gradle.kts
@@ -48,7 +48,6 @@ dependencies {
   implementation(projects.coreResources)
   implementation(projects.coreUi)
   implementation(projects.dataChatReadTimestampPublic)
-  implementation(projects.dataTravelCertificatePublic)
   implementation(projects.featureFlagsPublic)
   implementation(projects.languageCore)
   implementation(projects.marketCore)
@@ -73,7 +72,6 @@ dependencies {
   testImplementation(projects.apolloTest)
   testImplementation(projects.coreCommonTest)
   testImplementation(projects.dataChatReadTimestampTest)
-  testImplementation(projects.dataTravelCertificateTest)
   testImplementation(projects.featureFlagsTest)
   testImplementation(projects.languageTest)
   testImplementation(projects.loggingTest)
