+++ app/feature/feature-payments/src/main/kotlin/com/hedvig/android/feature/payments/navigation/PaymentsGraph.kt
@@ -96,8 +96,7 @@ fun NavGraphBuilder.paymentsGraph(
       )
     }
 
-    composable<PaymentsDestinations.Forever>(
-    ) {
+    composable<PaymentsDestinations.Forever> {
       val viewModel: ForeverViewModel = koinViewModel()
       ForeverDestination(
         viewModel = viewModel,
