+++ app/feature/feature-profile/src/main/kotlin/com/hedvig/android/feature/profile/tab/ProfileViewModel.kt
@@ -22,7 +22,7 @@ import kotlinx.coroutines.launch
 internal class ProfileViewModel(
   private val getEuroBonusStatusUseCase: GetEurobonusStatusUseCase,
   private val checkTravelCertificateDestinationAvailabilityUseCase:
-  CheckTravelCertificateDestinationAvailabilityUseCase,
+    CheckTravelCertificateDestinationAvailabilityUseCase,
   private val getMemberRemindersUseCase: GetMemberRemindersUseCase,
   private val enableNotificationsReminderManager: EnableNotificationsReminderManager,
   private val featureManager: FeatureManager,
