+++ app/feature/feature-profile/src/test/kotlin/com/hedvig/android/feature/profile/tab/ProfileViewModelTest.kt
@@ -381,7 +381,7 @@ private class FakeGetEurobonusStatusUseCase() : GetEurobonusStatusUseCase {
 }
 
 private class FakeCheckTravelCertificateDestinationAvailabilityUseCase :
-    CheckTravelCertificateDestinationAvailabilityUseCase {
+  CheckTravelCertificateDestinationAvailabilityUseCase {
   val turbine = Turbine<Either<TravelCertificateAvailabilityError, Unit>>(
     name = "TravelCertificateAvailability response",
   )
