+++ app/feature/feature-travel-certificate/src/main/kotlin/com/hedvig/android/feature/travelcertificate/di/TravelCertificateModule.kt
@@ -2,12 +2,17 @@ package com.hedvig.android.feature.travelcertificate.di
 
 import com.apollographql.apollo3.ApolloClient
 import com.hedvig.android.core.fileupload.DownloadPdfUseCase
-import com.hedvig.android.data.travelcertificate.CheckTravelCertificateAvailabilityForCurrentContractsUseCase
-import com.hedvig.android.data.travelcertificate.GetCoInsuredForContractUseCase
-import com.hedvig.android.data.travelcertificate.GetEligibleContractsWithAddressUseCase
-import com.hedvig.android.data.travelcertificate.GetTravelCertificateSpecificationsUseCase
-import com.hedvig.android.data.travelcertificate.GetTravelCertificatesHistoryUseCase
+import com.hedvig.android.feature.travelcertificate.data.CheckTravelCertificateAvailabilityForCurrentContractsUseCase
+import com.hedvig.android.feature.travelcertificate.data.CheckTravelCertificateAvailabilityForCurrentContractsUseCaseImpl
 import com.hedvig.android.feature.travelcertificate.data.CreateTravelCertificateUseCase
+import com.hedvig.android.feature.travelcertificate.data.GetCoInsuredForContractUseCase
+import com.hedvig.android.feature.travelcertificate.data.GetCoInsuredForContractUseCaseImpl
+import com.hedvig.android.feature.travelcertificate.data.GetEligibleContractsWithAddressUseCase
+import com.hedvig.android.feature.travelcertificate.data.GetEligibleContractsWithAddressUseCaseImpl
+import com.hedvig.android.feature.travelcertificate.data.GetTravelCertificateSpecificationsUseCase
+import com.hedvig.android.feature.travelcertificate.data.GetTravelCertificateSpecificationsUseCaseImpl
+import com.hedvig.android.feature.travelcertificate.data.GetTravelCertificatesHistoryUseCase
+import com.hedvig.android.feature.travelcertificate.data.GetTravelCertificatesHistoryUseCaseImpl
 import com.hedvig.android.feature.travelcertificate.navigation.TravelCertificateDestination
 import com.hedvig.android.feature.travelcertificate.ui.choose.ChooseContractForCertificateViewModel
 import com.hedvig.android.feature.travelcertificate.ui.generatewhen.TravelCertificateDateInputViewModel
@@ -15,6 +20,7 @@ import com.hedvig.android.feature.travelcertificate.ui.generatewho.TravelCertifi
 import com.hedvig.android.feature.travelcertificate.ui.history.CertificateHistoryViewModel
 import com.hedvig.android.feature.travelcertificate.ui.overview.TravelCertificateOverviewViewModel
 import com.hedvig.android.language.LanguageService
+import kotlinx.datetime.Clock
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
 
@@ -52,5 +58,23 @@ val travelCertificateModule = module {
       get<CreateTravelCertificateUseCase>(),
       get<GetCoInsuredForContractUseCase>(),
     )
+  }
+  single<GetTravelCertificateSpecificationsUseCase> {
+    GetTravelCertificateSpecificationsUseCaseImpl(get<ApolloClient>())
+  }
+  single<GetTravelCertificatesHistoryUseCase> {
+    GetTravelCertificatesHistoryUseCaseImpl(
+      get<ApolloClient>(),
+      get<Clock>(),
+    )
+  }
+  single<CheckTravelCertificateAvailabilityForCurrentContractsUseCase> {
+    CheckTravelCertificateAvailabilityForCurrentContractsUseCaseImpl(get<ApolloClient>())
+  }
+  single<GetEligibleContractsWithAddressUseCase> {
+    GetEligibleContractsWithAddressUseCaseImpl(get<ApolloClient>())
+  }
+  single<GetCoInsuredForContractUseCase> {
+    GetCoInsuredForContractUseCaseImpl(get<ApolloClient>())
   }
 }
