+++ app/feature/feature-travel-certificate/src/test/kotlin/com/hedvig/android/feature/travelcertificate/GetTravelCertificateSpecificationsUseCaseTest.kt
@@ -1,4 +1,4 @@
-package com.hedvig.android.data.travelcertificate
+package com.hedvig.android.feature.travelcertificate
 
 import assertk.assertThat
 import assertk.assertions.isEqualTo
@@ -38,22 +38,26 @@ internal class GetTravelCertificateSpecificationsUseCaseTest {
 
   @Test
   fun `when the feature flag is on and the network request fails, we get not Error response`() = runTest {
-    val travelCertificateUseCase = GetTravelCertificateSpecificationsUseCaseImpl(
-      apolloClient,
-    )
+    val travelCertificateUseCase =
+      com.hedvig.android.feature.travelcertificate.data.GetTravelCertificateSpecificationsUseCaseImpl(
+        apolloClient,
+      )
 
     apolloClient.enqueueTestNetworkError()
     val result = travelCertificateUseCase.invoke(null)
 
-    assertThat(result).isLeft().isInstanceOf<TravelCertificateError.Error>()
+    assertThat(
+      result,
+    ).isLeft().isInstanceOf<com.hedvig.android.feature.travelcertificate.data.TravelCertificateError.Error>()
   }
 
   @Test
   fun `when the feature flag is on and the network response contains no travel certificate, we get not eligible`() =
     runTest {
-      val travelCertificateUseCase = GetTravelCertificateSpecificationsUseCaseImpl(
-        apolloClient,
-      )
+      val travelCertificateUseCase =
+        com.hedvig.android.feature.travelcertificate.data.GetTravelCertificateSpecificationsUseCaseImpl(
+          apolloClient,
+        )
 
       apolloClient.enqueueTestResponse(
         TravelCertificateSpecificationsQuery(),
@@ -61,14 +65,17 @@ internal class GetTravelCertificateSpecificationsUseCaseTest {
       )
       val result = travelCertificateUseCase.invoke(null)
 
-      assertThat(result).isLeft().isInstanceOf<TravelCertificateError.NotEligible>()
+      assertThat(
+        result,
+      ).isLeft().isInstanceOf<com.hedvig.android.feature.travelcertificate.data.TravelCertificateError.NotEligible>()
     }
 
   @Test
   fun `when the passed contractId is wrong, we get not eligible`() = runTest {
-    val travelCertificateUseCase = GetTravelCertificateSpecificationsUseCaseImpl(
-      apolloClient,
-    )
+    val travelCertificateUseCase =
+      com.hedvig.android.feature.travelcertificate.data.GetTravelCertificateSpecificationsUseCaseImpl(
+        apolloClient,
+      )
 
     apolloClient.enqueueTestResponse(
       TravelCertificateSpecificationsQuery(),
@@ -76,14 +83,17 @@ internal class GetTravelCertificateSpecificationsUseCaseTest {
     )
     val result = travelCertificateUseCase.invoke("wrong contractId")
 
-    assertThat(result).isLeft().isInstanceOf<TravelCertificateError.NotEligible>()
+    assertThat(
+      result,
+    ).isLeft().isInstanceOf<com.hedvig.android.feature.travelcertificate.data.TravelCertificateError.NotEligible>()
   }
 
   @Test
   fun `when the passed contractId is right, we get TravelCertificateData`() = runTest {
-    val travelCertificateUseCase = GetTravelCertificateSpecificationsUseCaseImpl(
-      apolloClient,
-    )
+    val travelCertificateUseCase =
+      com.hedvig.android.feature.travelcertificate.data.GetTravelCertificateSpecificationsUseCaseImpl(
+        apolloClient,
+      )
 
     apolloClient.enqueueTestResponse(
       TravelCertificateSpecificationsQuery(),
@@ -107,8 +117,8 @@ internal class GetTravelCertificateSpecificationsUseCaseTest {
     val result = travelCertificateUseCase.invoke("id")
 
     assertThat(result).isRight().isEqualTo(
-      TravelCertificateData(
-        TravelCertificateData.TravelCertificateSpecification(
+      com.hedvig.android.feature.travelcertificate.data.TravelCertificateData(
+        com.hedvig.android.feature.travelcertificate.data.TravelCertificateData.TravelCertificateSpecification(
           contractId = "id",
           email = "email",
           maxDurationDays = 1,
@@ -121,9 +131,10 @@ internal class GetTravelCertificateSpecificationsUseCaseTest {
 
   @Test
   fun `when the passed contractId is correct, but the contract is not eligible, we get not eligible`() = runTest {
-    val travelCertificateUseCase = GetTravelCertificateSpecificationsUseCaseImpl(
-      apolloClient,
-    )
+    val travelCertificateUseCase =
+      com.hedvig.android.feature.travelcertificate.data.GetTravelCertificateSpecificationsUseCaseImpl(
+        apolloClient,
+      )
 
     apolloClient.enqueueTestResponse(
       TravelCertificateSpecificationsQuery(),
@@ -146,16 +157,19 @@ internal class GetTravelCertificateSpecificationsUseCaseTest {
     )
     val result = travelCertificateUseCase.invoke("id not eligible")
 
-    assertThat(result).isLeft().isInstanceOf<TravelCertificateError.NotEligible>()
+    assertThat(
+      result,
+    ).isLeft().isInstanceOf<com.hedvig.android.feature.travelcertificate.data.TravelCertificateError.NotEligible>()
   }
 
   @Test
   fun `when the feature flag is on and the network request succeeds, the response depends on the active contract travel certificate eligibility`(
     @TestParameter contractSupportsTravelCertificate: Boolean,
   ) = runTest {
-    val travelCertificateUseCase = GetTravelCertificateSpecificationsUseCaseImpl(
-      apolloClient,
-    )
+    val travelCertificateUseCase =
+      com.hedvig.android.feature.travelcertificate.data.GetTravelCertificateSpecificationsUseCaseImpl(
+        apolloClient,
+      )
 
     apolloClient.enqueueTestResponse(
       TravelCertificateSpecificationsQuery(),
@@ -186,8 +200,8 @@ internal class GetTravelCertificateSpecificationsUseCaseTest {
 
     if (contractSupportsTravelCertificate) {
       assertThat(result).isRight().isEqualTo(
-        TravelCertificateData(
-          TravelCertificateData.TravelCertificateSpecification(
+        com.hedvig.android.feature.travelcertificate.data.TravelCertificateData(
+          com.hedvig.android.feature.travelcertificate.data.TravelCertificateData.TravelCertificateSpecification(
             contractId = "contractId",
             email = "email",
             maxDurationDays = 1,
@@ -197,7 +211,9 @@ internal class GetTravelCertificateSpecificationsUseCaseTest {
         ),
       )
     } else {
-      assertThat(result).isLeft().isInstanceOf<TravelCertificateError.NotEligible>()
+      assertThat(
+        result,
+      ).isLeft().isInstanceOf<com.hedvig.android.feature.travelcertificate.data.TravelCertificateError.NotEligible>()
     }
   }
 }
